+++ src/action-delegate.ts
@@ -1,0 +1,3 @@
+"use client";
+export { ActionDelegate } from "./primitives/action-delegate/index.js";
+export type { ActionDelegateProps } from "./primitives/action-delegate/index.js";